requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
openai>=1.50.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
